requirements.txt
Django>=1.9.7,<2.0.0
django-tables2==1.2.3
djangorestframework==3.3.3
ipython==4.2.1
pysha3==0.3
django-s3-folder-storage==0.3
django-storages==1.4.1
boto==2.41.0
env-excavator==1.5.0
python-dotenv==0.5.1
psycopg2==2.6.2
dj-database-url==0.4.1
django-debug-toolbar==1.4
gunicorn==19.6.0
djangorestframework-filters==0.8.0
